requirements.txt
Flask
gunicorn
setuptools
python-dotenv
yfinance
pandas
pandas-ta
google-generativeai
requests
